Splits `crbot.go` into `crbot.go` and `system.go`
`system.go` contains all of the domain-specific logic

`crbot.go` contains everything needed to hook the system up with the outside
world.

This isn't a 100% split yet; there still needs to be interfaces for Gist and
Discord.

This is a move towards integration testing; once all of the external logic has
been abstracted behind interfaces, then it will be possible to create a large
integration test suite that creates fakes and passes them into everything in
`system.go`.