Refactors Parsers into their own interface.
- Changes Help commands to use the Name_* strings instead of types
- Changes the Feature interface to have Parsers() and FallbackParser()
- Removes the 1:1 mapping between Name_* and Type_*
- Feature registration now handles the fallback parser registration
- Help text now lives on parsers

Previously, it was assumed that executors and parsers had a 1:1 mapping, but
this is obviously not true (since many features emit Help commands). So this is
step 1 of decoupling parsers and executors. Next, executors will be refactored
into their own interface, and then finally, features will register many parsers
and executors. In this world, LearnFeature, UnlearnFeature, and CustomFeature
will all live together (since they all represent the same functionality).