Initial version of the ModelHelper, and vote features
The ModelHelper is a business layer on top of the string map, which allows
different operations to be performed and written against votes in the underlying
string map.

It is not complete; In a followup, the system clock will be abstracted such that
tests can advance time. Other followups will contain the commands and the timer
implementation.