Splits the util package into testutil and util
Testutil is used to store testing fakes, assertions for testing, and in the
future, the TestRunner

Util contains cross-project utility methods. For instance, one that collapses
whitespace

This is being done in order to allow TestRunner to be moved to testutil. Not
breaking this up previously caused an import loop between testutil and feature
(which relied on a non-testing util)