Moves the tests in the vote package into a parallel tests/ directory
This allows tests to work for integration testing with little hassle. The
integration testing framework that I wrote pulls in all the features, which does
not cause circular deps in Bazel (due to being able to define granularly what is
in each compilation unit), but it doesn't allow `go test` to work with tests in
the same directory.

All of the other packages will be ported over in the future, and `go test` added
to a prerelease script that I will run before pushing anything.