Moves all other tests into the tests/ directory
This is being done because I've taken an approach where a single test runner
depends on all features in order to make assertions on specific things within
them. This is creating circular package dependencies, which the Bazel build
allows but the Go tools themselves choke on.

In the future, this will be changed such that each feature defines its own test
runner that encapsulates the core test runner. Until then, this separation will
need to continue.